Move alert_words views into their own file.
