+++ zproject/urls.py
@@ -139,7 +139,7 @@ urlpatterns += patterns('zerver.views',
     url(r'^json/update_message$',           'messages.json_update_message'),
     url(r'^json/fetch_raw_message$',        'messages.json_fetch_raw_message'),
     url(r'^json/refer_friend$',             'json_refer_friend'),
-    url(r'^json/set_alert_words$',          'json_set_alert_words'),
+    url(r'^json/set_alert_words$',          'alert_words.json_set_alert_words'),
     url(r'^json/set_muted_topics$',         'json_set_muted_topics'),
     url(r'^json/set_avatar$',               'user_settings.json_set_avatar'),
     url(r'^json/time_setting$',             'user_settings.json_time_setting'),
@@ -184,11 +184,6 @@ v1_api_and_json_patterns = patterns('zerver.views',
     url(r'^users/me/pointer$', 'rest_dispatch',
             {'GET': 'get_pointer_backend',
              'PUT': 'update_pointer_backend'}),
-    url(r'^users/me/alert_words$', 'rest_dispatch',
-            {'GET': 'list_alert_words',
-             'PUT': 'set_alert_words',
-             'PATCH': 'add_alert_words',
-             'DELETE': 'remove_alert_words'}),
     url(r'^realm$', 'rest_dispatch',
             {'PATCH': 'update_realm'}),
     url(r'^users/me/presence$', 'rest_dispatch',
@@ -230,6 +225,13 @@ v1_api_and_json_patterns = patterns('zerver.views',
             {'GET':  'render_message_backend'}),
     url(r'^messages/flags$', 'rest_dispatch',
             {'POST':  'update_message_flags'}),
+
+) + patterns('zerver.views.alert_words',
+    url(r'^users/me/alert_words$', 'rest_dispatch',
+        {'GET': 'list_alert_words',
+         'PUT': 'set_alert_words',
+         'PATCH': 'add_alert_words',
+         'DELETE': 'remove_alert_words'}),
 
 ) + patterns('zerver.views.user_settings',
     url(r'^users/me/api_key/regenerate$', 'rest_dispatch',
